feat: from int to long

+++ common/src/main/java/by/kladvirov/dto/payment/json/Info.java
@@ -9,7 +9,7 @@ import lombok.Data;
 @Builder
 public class Info {
 
-    private Integer hours;
+    private Long hours;
 
     private Service service;
